% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{apply_alias}
\alias{apply_alias}
\title{clean up columns names}
\usage{
apply_alias(l)
}
\arguments{
\item{l}{list of strings representing column names}
}
\value{
list
}
\description{
clean up columns names
}
\keyword{PID}
\keyword{load}
