% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{cleanup_column_names_tbl}
\alias{cleanup_column_names_tbl}
\title{wrapper function for drop_column_dt and apply_alias}
\usage{
cleanup_column_names_tbl(PID)
}
\arguments{
\item{PID}{string}
}
\value{
NONE side effect
}
\description{
helper function
}
\keyword{PID}
\keyword{load}
