% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{create_meta_directory_csv}
\alias{create_meta_directory_csv}
\title{Unzip metadata CSV files}
\usage{
create_meta_directory_csv(zip.files.path, meta.files.path)
}
\arguments{
\item{zip.files.path}{string representing where zip files are stored}

\item{meta.files.path}{string representing where metadata csv files will be stored}
}
\value{
NONE (side effect function)
}
\description{
Unzip metadata CSV files from ZIP files into a folder. The latter is created if it does not exists.
}
\keyword{CSV}
\keyword{directory}
\keyword{metadata}
