% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{drop_columns_dt}
\alias{drop_columns_dt}
\title{drop up columns}
\usage{
drop_columns_dt(dt)
}
\arguments{
\item{dt}{data.frame}
}
\value{
data.frame
}
\description{
helper function
}
\keyword{PID}
\keyword{load}
