% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_available_field_combinations}
\alias{get_available_field_combinations}
\title{wrapper function for drop_column_dt and apply_alias}
\usage{
get_available_field_combinations(PID, fields)
}
\arguments{
\item{PID}{string}
}
\value{
data.frame
}
\description{
helper function
}
\keyword{PID}
\keyword{load}
