% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_api.R
\name{get_cansim_cube_metadata}
\alias{get_cansim_cube_metadata}
\title{Retrieve metadata for specified CANSIM table}
\usage{
get_cansim_cube_metadata(PID)
}
\arguments{
\item{PID}{a string representing the Product ID}
}
\value{
A tibble with metadata
}
\description{
Allows for the retrieval of metadatadata from Statistics Canada APIs
}
