% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_common_dim}
\alias{get_common_dim}
\title{returns list of  common dimension for a given list of PIDs}
\usage{
get_common_dim(l, init = NULL)
}
\arguments{
\item{l}{list of PIDs}

\item{init}{initial dimenstions to compare with dimensions from PIDs provided (default = NULL)}
}
\value{
list
}
\description{
returns list of  common dimension for a given list of PIDs
}
\examples{
get_common(c("14100370","14100367","14100366"))
}
\keyword{cansim}
\keyword{dimension}
\keyword{metadata}
