% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_different_dim}
\alias{get_different_dim}
\title{returns a title for a given PIDs
#'}
\usage{
get_different_dim(l)
}
\arguments{
\item{l}{list of strings representing PIDs}
}
\value{
lsit
}
\description{
returns a title for a given PIDs
#'
}
\keyword{PID}
\keyword{descrption}
\keyword{different}
