% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_dim_fields}
\alias{get_dim_fields}
\title{returns the fields present in a dimension for a given PID}
\usage{
get_dim_fields(PID, Dim)
}
\arguments{
\item{PID}{string one pid}

\item{Dim}{string list of common dimensions across PIDs that we want inspected}
}
\value{
lists of list
}
\description{
returns the fields present in a dimension for a given PID
}
\keyword{PID}
\keyword{dimension}
\keyword{field}
