% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_dimension}
\alias{get_dimension}
\title{returns the dimension for a given PID}
\usage{
get_dimension(dt)
}
\arguments{
\item{PID}{string one pid}

\item{Dim}{string list of common dimensions across PIDs that we want inspected}
}
\value{
lists of list
}
\description{
returns the dimension for a given PID
}
\keyword{PID}
\keyword{dimension}
\keyword{field}
