% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_highest_level_fields_in_dimension}
\alias{get_highest_level_fields_in_dimension}
\title{guess the highest field in a dimension}
\usage{
get_highest_level_fields_in_dimension(dt, dimension)
}
\arguments{
\item{dt}{string list of PIDs}

\item{dim}{string list of common dimensions across PIDs that we want inspected}
}
\value{
lists of list
}
\description{
**this is a guess and could be incorrect**. we usually get the user to validate the guess
}
\keyword{PID}
\keyword{dimension}
\keyword{field}
