% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_object_rCANSIM}
\alias{get_object_rCANSIM}
\title{get object from rCANSIM internal environment}
\usage{
get_object_rCANSIM(object_name)
}
\arguments{
\item{object_name}{string}
}
\value{
object
}
\description{
helper function
}
\keyword{environment}
\keyword{internal}
