% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_PID_description}
\alias{get_PID_description}
\title{returns a title for a given PIDs
#'}
\usage{
get_PID_description(PID)
}
\arguments{
\item{PID}{list of strings}
}
\value{
data.table
}
\description{
returns a title for a given PIDs
#'
}
\keyword{PID}
\keyword{descrption}
\keyword{title}
