% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{get_PID_with_dimensions}
\alias{get_PID_with_dimensions}
\title{returns list of PIDS containing specific fields}
\usage{
get_PID_with_dimensions(PID_list, dimensions)
}
\arguments{
\item{PID_list}{list of PIDs}

\item{dimensons}{list of Dimensions}
}
\value{
list of PIDs
}
\description{
returns list of PIDS containing specific fields
}
\examples{
test_PID <- c("14100370","14100367","14100366")
get_PID_with_fields(test_PID,c("Age group","Geography"))
}
\keyword{PID}
\keyword{dimension}
