% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{inspect_fields}
\alias{inspect_fields}
\title{inspect the fields in PID that share the same dimension}
\usage{
inspect_fields(list_PID, dim)
}
\arguments{
\item{dim}{string list of common dimensions across PIDs that we want inspected}

\item{list_dt}{string list of PIDs}
}
\value{
lists of list
}
\description{
function returns the fields that are common and different as separate list
}
\examples{
list_dt <- c("14100370","14100367")
common_dim <- get_common_dim(list_dt)
inspect_fields(list_dt,common_dim)
}
\keyword{PID}
\keyword{dimension}
\keyword{field}
\keyword{inspect}
