% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_api.R
\docType{data}
\name{list.of.labour.cansim_tbl}
\alias{list.of.labour.cansim_tbl}
\title{List of labour related CANSIM tables}
\format{An object of class \code{numeric} of length 172.}
\usage{
list.of.labour.cansim_tbl
}
\value{
list
}
\description{
convenience function
}
\keyword{cansim}
