% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{list_meta}
\alias{list_meta}
\title{returns list of dimension in a given table}
\usage{
list_meta(PID)
}
\arguments{
\item{PID}{String or list of strings}
}
\value{
list
}
\description{
returns list of dimension in a given table
}
\examples{
list_meta("14100370")
list_meta(c("14100370","14100367","14100366"))
}
\keyword{cansim}
\keyword{dimension}
\keyword{metadata}
