% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{load_table}
\alias{load_table}
\title{load actual data in memory}
\usage{
load_table(PID, path)
}
\arguments{
\item{PID}{list of strings representing PIDs}

\item{path}{string where csv files are stored}
}
\value{
data.table
}
\description{
load actual data in memory
}
\keyword{PID}
\keyword{load}
