% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{ls_environment}
\alias{ls_environment}
\title{Initializes the meta querying system}
\usage{
ls_environment()
}
\arguments{
\item{NONE}{side effect}
}
\value{
list
}
\description{
read all the nessary metadata present in the provided path for the querying system
}
\keyword{cansim}
\keyword{directory}
\keyword{metadata}
\keyword{set}
