% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{monthly}
\alias{monthly}
\title{return meta table filtered for monthly freq}
\usage{
monthly()
}
\arguments{
\item{None}{Use internal enviroment}
}
\value{
data.table
}
\description{
helper function
}
\keyword{meta}
