% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_api.R
\name{post_with_timeout_retry}
\alias{post_with_timeout_retry}
\title{get POST repsonce from Statistic Canada web API}
\usage{
post_with_timeout_retry(url, body, timeout = 200, retry = 3)
}
\arguments{
\item{url}{string}

\item{body}{string}

\item{timeout}{integer (in seconds) (default = 200)}

\item{retry}{integer (default = 3)}
}
\value{
POST response
}
\description{
get POST repsonce from Statistic Canada web API
}
\keyword{API}
\keyword{internal}
