% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{quarterly}
\alias{quarterly}
\title{return meta table filtered for quarterly freq}
\usage{
quarterly()
}
\arguments{
\item{None}{Use internal enviroment}
}
\value{
data.table
}
\description{
helper function
}
\keyword{meta}
