% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{set_meta_directory}
\alias{set_meta_directory}
\title{Initializes the meta querying system}
\usage{
set_meta_directory(method, ...)
}
\arguments{
\item{method}{String (Values = c("API", "CSV")}
}
\value{
NONE (side effect function)
}
\description{
read all the nessary metadata present in the provided path for the querying system
}
\details{
The api method takes a list of CANSIM PIDs are extra argument

The CSV method take a string representing the path where the metadata CSVs are stored
}
\keyword{cansim}
\keyword{directory}
\keyword{internal}
\keyword{metadata}
\keyword{set}
