% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{set_meta_directory_api}
\alias{set_meta_directory_api}
\title{Initializes the meta querying system using Statistics Canada API}
\usage{
set_meta_directory_api(PID_List)
}
\arguments{
\item{PID_List}{List of strings representing that tables PIDs}
}
\value{
NONE (side effect function)
}
\description{
read all the nessary metadata present in the provided path for the querying system
}
\keyword{cansim}
\keyword{directory}
\keyword{internal}
\keyword{metadata}
\keyword{set}
