% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{set_meta_directory_csv}
\alias{set_meta_directory_csv}
\title{Initializes the meta querying system using Statistics Canada API}
\usage{
set_meta_directory_csv(path)
}
\arguments{
\item{path}{string representing where metadata csv files are stored}
}
\value{
NONE (side effect function)
}
\description{
read all the nessary metadata present in the provided path for the querying system
}
\keyword{cansim}
\keyword{directory}
\keyword{internal}
\keyword{metadata}
\keyword{set}
