% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_merge_pairwise.R
\name{smart_merge_pairwise}
\alias{smart_merge_pairwise}
\title{interactive merging of cansim tables}
\usage{
smart_merge_pairwise(dt, dt2, VERBOSE = TRUE)
}
\arguments{
\item{dt}{String representing PID}

\item{dt2}{String representing PID}

\item{VERBOSE}{Boolean to trigger the extra display of information for the merge}
}
\value{
data.table
}
\description{
interactive merging of cansim tables ( more to be added later)
}
\keyword{cansim}
\keyword{merge}
