% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{yearly}
\alias{yearly}
\title{return meta table filtered for yearly freq}
\usage{
yearly()
}
\arguments{
\item{None}{Use internal enviroment}
}
\value{
data.table
}
\description{
helper function
}
\keyword{meta}
